requirements.txt
python-telegram-bot==21.0
Flask==3.0.0
gunicorn==21.2.0
APScheduler==3.10.4
psutil==5.9.6
python-dotenv==1.0.0
